extensions tidy up
